+++ zebedee-cms/src/test/java/com/github/onsdigital/zebedee/keyring/legacy/BaseLegacyKeyringTest.java
@@ -124,6 +124,9 @@ public abstract class BaseLegacyKeyringTest {
         when(collection.getDescription())
                 .thenReturn(collectionDescription);
 
+        when(collection.getId())
+                .thenReturn(TEST_COLLECTION_ID);
+
         when(collectionDescription.getId())
                 .thenReturn(TEST_COLLECTION_ID);
