+++ zebedee-cms/src/test/java/com/github/onsdigital/zebedee/keyring/migration/MigrationCollectionKeyringImplTest.java
@@ -42,26 +42,26 @@ import static org.mockito.Mockito.when;
 public class MigrationCollectionKeyringImplTest {
 
     @Mock
-    private User user;
+    protected User user;
 
     @Mock
-    private Collection collection;
+    protected Collection collection;
 
     @Mock
-    private SecretKey secretKey;
+    protected SecretKey secretKey;
 
     @Mock
-    private CollectionKeyring legacyCollectionKeyring;
+    protected CollectionKeyring legacyCollectionKeyring;
 
     @Mock
-    private CollectionKeyring collectionKeyring;
+    protected CollectionKeyring collectionKeyring;
 
-    private CollectionKeyring keyring;
+    protected CollectionKeyring keyring;
 
-    private KeyringException keyringException;
+    protected KeyringException keyringException;
 
-    private boolean enabled = true;
-    private boolean disabled = false;
+    protected boolean enabled = true;
+    protected boolean disabled = false;
 
     @Before
     public void setUp() throws Exception {
@@ -70,815 +70,14 @@ public class MigrationCollectionKeyringImplTest {
         this.keyringException = new KeyringException("keyringException");
     }
 
-    private void assertWrappedException(KeyringException actual, String expectedMsg, boolean migrationEnabled) {
+    protected void assertWrappedException(KeyringException actual, String expectedMsg, boolean migrationEnabled) {
         String msg = format(WRAPPED_ERR_FMT, expectedMsg, migrationEnabled);
         assertThat(actual.getMessage(), equalTo(msg));
         assertThat(actual.getCause(), is(notNullValue()));
     }
 
-    private void assertException(KeyringException actual, String expectedMsg, boolean migrationEnabled) {
+    protected void assertException(KeyringException actual, String expectedMsg, boolean migrationEnabled) {
         String msg = format(WRAPPED_ERR_FMT, expectedMsg, migrationEnabled);
         assertThat(actual.getMessage(), equalTo(msg));
-    }
-
-    @Test
-    public void cacheKeyring_migrateDisabled_success_shouldUpdateBothKeyrings() throws Exception {
-        // Given keyring migration is disabled
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        // When cacheKeyring is called
-        keyring.cacheKeyring(user);
-
-        // Then both the legacy and central keyrings are updated
-        verify(legacyCollectionKeyring, times(1)).cacheKeyring(user);
-        verify(collectionKeyring, times(1)).cacheKeyring(user);
-    }
-
-    @Test
-    public void cacheKeyring_migrateEnabled_success_shouldUpdateBothKeyrings() throws Exception {
-        // Given keyring migration is enabled
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        // When cacheKeyring is called
-        keyring.cacheKeyring(user);
-
-        // Then both the legacy and central keyrings are updated
-        verify(legacyCollectionKeyring, times(1)).cacheKeyring(user);
-        verify(collectionKeyring, times(1)).cacheKeyring(user);
-    }
-
-    @Test
-    public void cacheKeyring_migrateDisabled_legacyKeyringException_shouldThrowException() throws Exception {
-        // Given keyring migration is disabled
-        // And legacy keyring errors when called.
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        doThrow(KeyringException.class)
-                .when(legacyCollectionKeyring)
-                .cacheKeyring(user);
-
-        // When cacheKeyring is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.cacheKeyring(user));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, POPULATE_FROM_USER_ERR, disabled);
-
-        // And the legacy keyring is called 1 time
-        verify(legacyCollectionKeyring, times(1)).cacheKeyring(user);
-
-        // And the central keyring is never called
-        verifyZeroInteractions(collectionKeyring);
-    }
-
-    @Test
-    public void cacheKeyring_migrateEnabled_legacyKeyringException_shouldThrowException() throws Exception {
-        // Given keyring migration is enabled
-        // And legacy keyring errors when called.
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        doThrow(KeyringException.class)
-                .when(legacyCollectionKeyring)
-                .cacheKeyring(user);
-
-        // When cacheKeyring is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.cacheKeyring(user));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, POPULATE_FROM_USER_ERR, enabled);
-
-        // And the legacy keyring is called 1 time
-        verify(legacyCollectionKeyring, times(1)).cacheKeyring(user);
-
-        // And the central keyring is never called
-        verifyZeroInteractions(collectionKeyring);
-    }
-
-    @Test
-    public void cacheKeyring_migrateDisabled_centralKeyringException_shouldThrowException() throws Exception {
-        // Given keyring migration is disabled
-        // And central keyring errors when called.
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        doThrow(KeyringException.class)
-                .when(collectionKeyring)
-                .cacheKeyring(user);
-
-        // When cacheKeyring is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.cacheKeyring(user));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, POPULATE_FROM_USER_ERR, disabled);
-
-        // And the legacy keyring is called 1 time
-        verify(legacyCollectionKeyring, times(1)).cacheKeyring(user);
-
-        // And the central keyring is called 1 time
-        verify(collectionKeyring, times(1)).cacheKeyring(user);
-    }
-
-    @Test
-    public void cacheKeyring_migrateEnabled_centralKeyringException_shouldThrowException() throws Exception {
-        // Given keyring migration is enabled
-        // And central keyring errors when called.
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        doThrow(KeyringException.class)
-                .when(collectionKeyring)
-                .cacheKeyring(user);
-
-        // When cacheKeyring is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.cacheKeyring(user));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, POPULATE_FROM_USER_ERR, enabled);
-
-        // And the legacy keyring is called 1 time
-        verify(legacyCollectionKeyring, times(1)).cacheKeyring(user);
-
-        // And the central keyring is called 1 time
-        verify(collectionKeyring, times(1)).cacheKeyring(user);
-    }
-
-    @Test
-    public void get_migrateDisabled_success_shouldReadFromLegacyKeyring() throws Exception {
-        // Given keyring migration is disabled
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        when(legacyCollectionKeyring.get(user, collection))
-                .thenReturn(secretKey);
-
-        // When Get is called
-        SecretKey actual = keyring.get(user, collection);
-
-        // Then the expected key is returned
-        assertThat(actual, equalTo(secretKey));
-        verify(legacyCollectionKeyring, times(1)).get(user, collection);
-        verifyZeroInteractions(collectionKeyring);
-    }
-
-    @Test
-    public void get_migrateDisabled_keyringError_shouldThrowException() throws Exception {
-        // Given keyring migration is disabled
-        // And legacy keyring returns an error
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        when(legacyCollectionKeyring.get(user, collection))
-                .thenThrow(KeyringException.class);
-
-        // When Get is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.get(user, collection));
-
-        // Then an exception is returned
-        assertWrappedException(actual, GET_KEY_ERR, disabled);
-
-        // And the legacy keyring is called 1 times
-        verify(legacyCollectionKeyring, times(1)).get(user, collection);
-
-        // And the central keyring is never called
-        verifyZeroInteractions(collectionKeyring);
-    }
-
-    @Test
-    public void get_migrateEnabled_success_shouldReadFromCentralKeyring() throws Exception {
-        // Given keyring migration is enabled
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        when(collectionKeyring.get(user, collection))
-                .thenReturn(secretKey);
-
-        // When Get is called
-        SecretKey actual = keyring.get(user, collection);
-
-        // Then the expected key is returned
-        assertThat(actual, equalTo(secretKey));
-        verify(collectionKeyring, times(1)).get(user, collection);
-        verifyZeroInteractions(legacyCollectionKeyring);
-    }
-
-    @Test
-    public void get_migrateEnabled_keyringError_shouldThrowException() throws Exception {
-        // Given keyring migration is enabled
-        // And central keyring returns an error
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        when(collectionKeyring.get(user, collection))
-                .thenThrow(KeyringException.class);
-
-        // When Get is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.get(user, collection));
-
-        // Then an exception is returned
-        assertWrappedException(actual, GET_KEY_ERR, enabled);
-
-        // And the central keyring is called 1 times
-        verify(collectionKeyring, times(1)).get(user, collection);
-
-        // And the legacy keyring is never called
-        verifyZeroInteractions(legacyCollectionKeyring);
-    }
-
-    @Test
-    public void remove_migrateDisabled_getKeyError_shouldThrownException() throws Exception {
-        // Given migration is disabled
-        // And get key returns an error
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        when(legacyCollectionKeyring.get(user, collection))
-                .thenThrow(keyringException);
-
-        // When remove is called
-        KeyringException ex = assertThrows(KeyringException.class, () -> keyring.remove(user, collection));
-
-        // Then an exception is thrown
-        assertWrappedException(ex, GET_KEY_ERR, disabled);
-        verify(legacyCollectionKeyring, times(1)).get(user, collection);
-        verify(legacyCollectionKeyring, never()).remove(user, collection);
-        verify(collectionKeyring, never()).remove(user, collection);
-    }
-
-    @Test
-    public void remove_migrateEnabled_getKeyError_shouldThrownException() throws Exception {
-        // Given migration is enabled
-        // And get key returns an error
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        when(collectionKeyring.get(user, collection))
-                .thenThrow(keyringException);
-
-        // When remove is called
-        KeyringException ex = assertThrows(KeyringException.class, () -> keyring.remove(user, collection));
-
-        // Then an exception is thrown
-        assertWrappedException(ex, GET_KEY_ERR, enabled);
-        verify(collectionKeyring, times(1)).get(user, collection);
-        verify(legacyCollectionKeyring, never()).remove(user, collection);
-        verify(legacyCollectionKeyring, never()).remove(user, collection);
-    }
-
-    @Test
-    public void remove_migrateDisabled_keyNull_shouldThrowException() throws Exception {
-        // Given migration is disabled
-        // And get key returns null
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        when(legacyCollectionKeyring.get(user, collection))
-                .thenReturn(null);
-
-        // When remove is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.remove(user, collection));
-
-        // Then an exception is thrown
-        assertException(actual, KEY_NULL_ERR, disabled);
-        verify(legacyCollectionKeyring, times(1)).get(user, collection);
-        verify(collectionKeyring, never()).remove(user, collection);
-        verify(collectionKeyring, never()).remove(user, collection);
-    }
-
-    @Test
-    public void remove_migrateEnabled_keyNull_shouldThrowException() throws Exception {
-        // Given migration is enabled
-        // And get key returns null
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        when(collectionKeyring.get(user, collection))
-                .thenReturn(null);
-
-        // When remove is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.remove(user, collection));
-
-        // Then an exception is thrown
-        assertException(actual, KEY_NULL_ERR, enabled);
-        verify(collectionKeyring, times(1)).get(user, collection);
-        verify(legacyCollectionKeyring, never()).remove(user, collection);
-        verify(legacyCollectionKeyring, never()).remove(user, collection);
-    }
-
-    @Test
-    public void remove_migrateDisabled_centralKeyringRemoveException_shouldThrowException() throws KeyringException {
-        // Given migration is disabled
-        // And central keyring returns an error
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        when(legacyCollectionKeyring.get(user, collection))
-                .thenReturn(secretKey);
-
-        doThrow(KeyringException.class)
-                .when(collectionKeyring)
-                .remove(user, collection);
-
-        // When remove is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.remove(user, collection));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, REMOVE_KEY_ERR, disabled);
-        verify(legacyCollectionKeyring, times(1)).get(user, collection);
-        verify(collectionKeyring, times(1)).remove(user, collection);
-        verifyNoMoreInteractions(legacyCollectionKeyring);
-    }
-
-    @Test
-    public void remove_migrateEnabled_centralKeyringRemoveException_shouldThrowException() throws KeyringException {
-        // Given migration is enabled
-        // And central keyring returns an error
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        when(collectionKeyring.get(user, collection))
-                .thenReturn(secretKey);
-
-        doThrow(KeyringException.class)
-                .when(collectionKeyring)
-                .remove(user, collection);
-
-        // When remove is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.remove(user, collection));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, REMOVE_KEY_ERR, enabled);
-        verify(collectionKeyring, times(1)).get(user, collection);
-        verify(collectionKeyring, times(1)).remove(user, collection);
-        verifyZeroInteractions(legacyCollectionKeyring);
-    }
-
-    @Test
-    public void remove_migrateDisabled_legacyKeyringRemoveException_shouldThrowException() throws KeyringException {
-        // Given migration is disabled
-        // And legacy keyring returns an error
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        when(legacyCollectionKeyring.get(user, collection))
-                .thenReturn(secretKey);
-
-        doThrow(KeyringException.class)
-                .when(legacyCollectionKeyring)
-                .remove(user, collection);
-
-        // When remove is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.remove(user, collection));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, REMOVE_KEY_ERR, disabled);
-        verify(legacyCollectionKeyring, times(1)).get(user, collection);
-        verify(legacyCollectionKeyring, times(1)).remove(user, collection);
-        verify(collectionKeyring, times(1)).add(user, collection, secretKey);
-    }
-
-    @Test
-    public void remove_migrateEnabled_legacyKeyringRemoveException_shouldThrowException() throws KeyringException {
-        // Given migration is enabled
-        // And legacy keyring returns an error
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        when(collectionKeyring.get(user, collection))
-                .thenReturn(secretKey);
-
-        doThrow(keyringException)
-                .when(legacyCollectionKeyring)
-                .remove(user, collection);
-
-        // When remove is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.remove(user, collection));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, REMOVE_KEY_ERR, enabled);
-        verify(collectionKeyring, times(1)).get(user, collection);
-        verify(collectionKeyring, times(1)).remove(user, collection);
-        verify(legacyCollectionKeyring, times(1)).remove(user, collection);
-        verify(collectionKeyring, times(1)).add(user, collection, secretKey);
-    }
-
-    @Test
-    public void remove_migrateDisabled_removeRollbackFails_shouldThrowException() throws KeyringException {
-        // Given migration is disabled
-        // And legacy keyring returns an error
-        // And remove rollback fails
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        when(legacyCollectionKeyring.get(user, collection))
-                .thenReturn(secretKey);
-
-        doThrow(keyringException)
-                .when(legacyCollectionKeyring)
-                .remove(user, collection);
-
-        doThrow(keyringException)
-                .when(collectionKeyring)
-                .add(user, collection, secretKey);
-
-        // When remove is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.remove(user, collection));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, ROLLBACK_FAILED_ERR, disabled);
-        verify(legacyCollectionKeyring, times(1)).get(user, collection);
-        verify(collectionKeyring, times(1)).remove(user, collection);
-        verify(legacyCollectionKeyring, times(1)).remove(user, collection);
-        verify(collectionKeyring, times(1)).add(user, collection, secretKey);
-    }
-
-    @Test
-    public void remove_migrateEnabled_removeRollbackFails_shouldThrowException() throws KeyringException {
-        // Given migration is enabled
-        // And legacy keyring returns an error
-        // And remove rollback fails
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        when(collectionKeyring.get(user, collection))
-                .thenReturn(secretKey);
-
-        doThrow(keyringException)
-                .when(legacyCollectionKeyring)
-                .remove(user, collection);
-
-        doThrow(keyringException)
-                .when(collectionKeyring)
-                .add(user, collection, secretKey);
-
-        // When remove is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.remove(user, collection));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, ROLLBACK_FAILED_ERR, enabled);
-        verify(collectionKeyring, times(1)).get(user, collection);
-        verify(collectionKeyring, times(1)).remove(user, collection);
-        verify(legacyCollectionKeyring, times(1)).remove(user, collection);
-        verify(collectionKeyring, times(1)).add(user, collection, secretKey);
-    }
-
-    @Test
-    public void remove_migrateDisabled_success_shouldRemoveKeyFromBothKeyrings() throws Exception {
-        // Given migration is disabled
-        // And remove is successful
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        when(legacyCollectionKeyring.get(user, collection))
-                .thenReturn(secretKey);
-
-        // When remove is called
-        keyring.remove(user, collection);
-
-        // Then both the legacy and central keyrings are updated
-        verify(legacyCollectionKeyring, times(1)).get(user, collection);
-        verify(collectionKeyring, times(1)).remove(user, collection);
-        verify(legacyCollectionKeyring, times(1)).remove(user, collection);
-        verify(collectionKeyring, never()).add(user, collection, secretKey);
-    }
-
-    @Test
-    public void remove_migrateEnabled_success_shouldRemoveKeyFromBothKeyrings() throws Exception {
-        // Given migration is enabled
-        // And remove is successful
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        when(collectionKeyring.get(user, collection))
-                .thenReturn(secretKey);
-
-        // When remove is called
-        keyring.remove(user, collection);
-
-        // Then both the legacy and central keyrings are updated
-        verify(collectionKeyring, times(1)).get(user, collection);
-        verify(collectionKeyring, times(1)).remove(user, collection);
-        verify(legacyCollectionKeyring, times(1)).remove(user, collection);
-        verify(collectionKeyring, never()).add(user, collection, secretKey);
-    }
-
-    @Test
-    public void add_migrateDisabled_legacyKeyringError_shouldThrowException() throws Exception {
-        // Given migration is disabled
-        // And legacy keyring.add is unsuccessful
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        doThrow(keyringException).
-                when(legacyCollectionKeyring)
-                .add(user, collection, secretKey);
-
-        // When add is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.add(user, collection, secretKey));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, ADD_KEY_ERR, disabled);
-
-        verify(legacyCollectionKeyring, times(1)).add(user, collection, secretKey);
-        verifyZeroInteractions(collectionKeyring);
-    }
-
-    @Test
-    public void add_migrateEnabled_legacyKeyringError_shouldThrowException() throws Exception {
-        // Given migration is enabled
-        // And legacy keyring.add is unsuccessful
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        doThrow(keyringException).
-                when(legacyCollectionKeyring)
-                .add(user, collection, secretKey);
-
-        // When add is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.add(user, collection, secretKey));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, ADD_KEY_ERR, enabled);
-
-        verify(legacyCollectionKeyring, times(1)).add(user, collection, secretKey);
-        verifyZeroInteractions(collectionKeyring);
-    }
-
-    @Test
-    public void add_migrateDisabled_rollbackUnsuccessful_shouldThrowException() throws Exception {
-        // Given migration is disabled
-        // And central keyring.add is unsuccessful
-        // And rollback is unsuccessful
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        doThrow(keyringException).
-                when(collectionKeyring)
-                .add(user, collection, secretKey);
-
-        doThrow(keyringException).
-                when(legacyCollectionKeyring)
-                .remove(user, collection);
-
-        // When add is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.add(user, collection, secretKey));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, ROLLBACK_FAILED_ERR, disabled);
-
-        verify(legacyCollectionKeyring, times(1)).add(user, collection, secretKey);
-        verify(collectionKeyring, times(1)).add(user, collection, secretKey);
-        verify(legacyCollectionKeyring, times(1)).remove(user, collection);
-    }
-
-    @Test
-    public void add_migrateEnabled_rollbackUnsuccessful_shouldThrowException() throws Exception {
-        // Given migration is enabled
-        // And central keyring.add is unsuccessful
-        // And rollback is unsuccessful
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        doThrow(keyringException).
-                when(collectionKeyring)
-                .add(user, collection, secretKey);
-
-        doThrow(keyringException).
-                when(legacyCollectionKeyring)
-                .remove(user, collection);
-
-        // When add is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.add(user, collection, secretKey));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, ROLLBACK_FAILED_ERR, enabled);
-
-        verify(legacyCollectionKeyring, times(1)).add(user, collection, secretKey);
-        verify(collectionKeyring, times(1)).add(user, collection, secretKey);
-        verify(legacyCollectionKeyring, times(1)).remove(user, collection);
-    }
-
-    @Test
-    public void add_migrateDisabled_rollbackSuccessful_shouldThrowException() throws Exception {
-        // Given migration is disabled
-        // And central keyring.add is unsuccessful
-        // And rollback is successful
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        doThrow(keyringException).
-                when(collectionKeyring)
-                .add(user, collection, secretKey);
-
-        // When add is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.add(user, collection, secretKey));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, ADD_KEY_ERR, disabled);
-
-        verify(legacyCollectionKeyring, times(1)).add(user, collection, secretKey);
-        verify(collectionKeyring, times(1)).add(user, collection, secretKey);
-        verify(legacyCollectionKeyring, times(1)).remove(user, collection);
-    }
-
-    @Test
-    public void add_migrateEnabled_rollbackSuccessful_shouldThrowException() throws Exception {
-        // Given migration is enabled
-        // And central keyring.add is unsuccessful
-        // And rollback is successful
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        doThrow(keyringException).
-                when(collectionKeyring)
-                .add(user, collection, secretKey);
-
-        // When add is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.add(user, collection, secretKey));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, ADD_KEY_ERR, enabled);
-
-        verify(legacyCollectionKeyring, times(1)).add(user, collection, secretKey);
-        verify(collectionKeyring, times(1)).add(user, collection, secretKey);
-        verify(legacyCollectionKeyring, times(1)).remove(user, collection);
-    }
-
-    @Test
-    public void add_migrateDisabled_success_shouldAddKeyToBothKeyrings() throws Exception {
-        // Given migration is disabled
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        // When add is called
-        keyring.add(user, collection, secretKey);
-
-        // Then the key is added to both keyrings
-        verify(legacyCollectionKeyring, times(1)).add(user, collection, secretKey);
-        verify(collectionKeyring, times(1)).add(user, collection, secretKey);
-        verifyNoMoreInteractions(legacyCollectionKeyring, collectionKeyring);
-    }
-
-    @Test
-    public void add_migrateEnabled_success_shouldAddKeyToBothKeyrings() throws Exception {
-        // Given migration is enabled
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        // When add is called
-        keyring.add(user, collection, secretKey);
-
-        // Then the key is added to both keyrings
-        verify(legacyCollectionKeyring, times(1)).add(user, collection, secretKey);
-        verify(collectionKeyring, times(1)).add(user, collection, secretKey);
-        verifyNoMoreInteractions(legacyCollectionKeyring, collectionKeyring);
-    }
-
-    @Test
-    public void list_migrateDisabled_listError_shouldThrowException() throws Exception {
-        // Given migrate is disabled
-        // And legacy keyring.list throws an exception
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        when(legacyCollectionKeyring.list(user))
-                .thenThrow(keyringException);
-
-        // When list is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.list(user));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, LIST_KEYS_ERR, disabled);
-        verify(legacyCollectionKeyring, times(1)).list(user);
-        verifyZeroInteractions(collectionKeyring);
-    }
-
-    @Test
-    public void list_migrateEnabled_listError_shouldThrowException() throws Exception {
-        // Given migrate is enabled
-        // And legacy keyring.list throws an exception
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        when(collectionKeyring.list(user))
-                .thenThrow(keyringException);
-
-        // When list is called
-        KeyringException actual = assertThrows(KeyringException.class, () -> keyring.list(user));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, LIST_KEYS_ERR, enabled);
-        verify(collectionKeyring, times(1)).list(user);
-        verifyZeroInteractions(legacyCollectionKeyring);
-    }
-
-    @Test
-    public void list_migrateDisabled_success_shouldListKeys() throws Exception {
-        // Given migrate is disabled
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        Set<String> keys = new HashSet<String>() {{
-            add("1234567890");
-        }};
-
-        when(legacyCollectionKeyring.list(user))
-                .thenReturn(keys);
-
-        // When list is called
-        Set<String> actual = keyring.list(user);
-
-        // Then the expected key list is returned
-        assertThat(actual, equalTo(keys));
-        verify(legacyCollectionKeyring, times(1)).list(user);
-        verifyZeroInteractions(collectionKeyring);
-    }
-
-    @Test
-    public void list_migrateEnabled_success_shouldListKeys() throws Exception {
-        // Given migrate is enabled
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        Set<String> keys = new HashSet<String>() {{
-            add("1234567890");
-        }};
-
-        when(collectionKeyring.list(user))
-                .thenReturn(keys);
-
-        // When list is called
-        Set<String> actual = keyring.list(user);
-
-        // Then the expected key list is returned
-        assertThat(actual, equalTo(keys));
-        verify(collectionKeyring, times(1)).list(user);
-        verifyZeroInteractions(legacyCollectionKeyring);
-    }
-
-    @Test
-    public void unlock_migrateDisbaled_shouldCallLegacyKeyringUnlock() throws Exception {
-        // Given migrate is disabled
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        // When unlocked is called
-        keyring.unlock(user, "1234");
-
-        // Then legacy keyring is unlocked
-        verify(legacyCollectionKeyring, times(1)).unlock(user, "1234");
-        verifyZeroInteractions(collectionKeyring);
-    }
-
-    @Test
-    public void unlock_migrateEnabled_shouldCallLegacyKeyringOnly() throws Exception {
-        // Given migrate is enabled
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        // When unlocked is called
-        keyring.unlock(user, "1234");
-
-        // Then central keyring is unlocked
-        verify(legacyCollectionKeyring, times(1)).unlock(user, "1234");
-        verifyZeroInteractions(collectionKeyring);
-    }
-
-    @Test
-    public void unlock_migrateDisabled_unlockErrorShouldThrowException() throws Exception {
-        // Given migrate is disabled
-        // And legacy keyring unlock throws an exception
-        keyring = new MigrationCollectionKeyringImpl(disabled, legacyCollectionKeyring, collectionKeyring);
-
-        String password = "1234";
-
-        doThrow(KeyringException.class)
-                .when(legacyCollectionKeyring)
-                .unlock(user, password);
-
-        // When unlocked is called
-        KeyringException actual = assertThrows(KeyringException.class, ()->keyring.unlock(user, password));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, UNLOCK_KEYRING_ERR, disabled);
-
-        verify(legacyCollectionKeyring, times(1)).unlock(user, password);
-        verifyZeroInteractions(collectionKeyring);
-    }
-
-    @Test
-    public void unlock_migrateEnabled_unlockErrorShouldThrowException() throws Exception {
-        // Given migrate is enabled
-        // And legacy keyring unlock throws an exception
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        String password = "1234";
-
-        doThrow(KeyringException.class)
-                .when(legacyCollectionKeyring)
-                .unlock(user, password);
-
-        // When unlocked is called
-        KeyringException actual = assertThrows(KeyringException.class, ()->keyring.unlock(user, password));
-
-        // Then an exception is thrown
-        assertWrappedException(actual, UNLOCK_KEYRING_ERR, enabled);
-
-        verify(legacyCollectionKeyring, times(1)).unlock(user, password);
-        verifyZeroInteractions(collectionKeyring);
-    }
-
-
-    @Test
-    public void testAssignTo_shouldCallLegacyKeyring() throws Exception {
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        User src = mock(User.class);
-        User target = mock(User.class);
-        List<CollectionDescription> assignments = mock(List.class);
-
-        keyring.assignTo(src, target, assignments);
-
-        verify(legacyCollectionKeyring, times(1)).assignTo(src, target, assignments);
-        verifyZeroInteractions(collectionKeyring);
-    }
-
-    @Test
-    public void testRemoveFrom_shouldCallLegacyKeyring() throws Exception {
-        keyring = new MigrationCollectionKeyringImpl(enabled, legacyCollectionKeyring, collectionKeyring);
-
-        User target = mock(User.class);
-        List<CollectionDescription> removals = mock(List.class);
-
-        keyring.revokeFrom(target, removals);
-
-        verify(legacyCollectionKeyring, times(1)).revokeFrom(target, removals);
-        verifyZeroInteractions(collectionKeyring);
     }
 }
